requirements.txt
alembic==1.3.1
aniso8601==8.0.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
autoenv==1.0.0
bcrypt==3.1.7
blinker==1.4
cffi==1.13.2
Click==7.0
coverage==4.5.4
Flask==1.1.1
flask-accepts==0.15.0
Flask-API==1.1
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.2
flask-requests==0.0.14
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
lazy-object-proxy==1.4.3
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.2
mccabe==0.6.1
more-itertools==7.2.0
mypy-extensions==0.4.3
packaging==19.2
pampy==0.3.0
pluggy==0.13.0
psycopg2==2.8.4
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.5
pyrsistent==0.15.5
pytest==5.2.2
pytest-cov==2.8.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.11
typed-ast==1.4.0
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
